% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_analysis.R
\name{coexpression_analysis}
\alias{coexpression_analysis}
\title{Calculate Gene Correlations}
\usage{
coexpression_analysis(expression_matrix, gene, method = "pearson")
}
\arguments{
\item{expression_matrix}{A numeric matrix where rows represent genes and columns represent samples.}

\item{gene}{A character string representing the gene for which the correlations will be calculated.}

\item{method}{A character string specifying the correlation method to be used. Default is "pearson".}
}
\value{
A data frame containing gene names, correlation coefficients, and p-values.
}
\description{
This function calculates the correlation between a given gene and all other genes
in the provided expression matrix. It also provides the corresponding p-values.
}
\examples{
\dontrun{
  expression_matrix <- get_OSF_data(table = "GSE74706", action = "geo_data")
  results <- coexpression_analysis(expression_matrix, "RPN1")
  print(results)
}
}
