% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_cancer_genelist.R
\name{cor_cancer_genelist}
\alias{cor_cancer_genelist}
\title{Perform correlation analysis}
\usage{
cor_cancer_genelist(
  dataset = "GSE62113",
  id1 = "STAT3",
  id2 = c("TNS1", "TP53"),
  tumor_subtype = NULL,
  sample_type = c("Tumor", "Normal"),
  cor_method = "pearson",
  adjust_method = "BH",
  conf_level = 0.95
)
}
\arguments{
\item{dataset}{Dataset name. Use 'dataset$Abbre' to get all datasets.}

\item{id1}{Gene symbol, you can input one gene symbol.}

\item{id2}{Gene symbols, you can input one or multiple symbols.}

\item{tumor_subtype}{Tumor subtype used for correlation analysis, default is NULL.}

\item{sample_type}{Sample type used for correlation analysis, default all types: c("Tumor", "Normal").}

\item{cor_method}{Method for correlation analysis, default "pearson".}

\item{adjust_method}{Method for adjusting p-values for multiple testing. Options include "none", "BH", "BY", "holm", "hochberg", "hommel", "bonferroni". Default is "BH".}

\item{conf_level}{The confidence level for confidence interval calculation (default: 0.95).}
}
\value{
A list containing the correlation results (including n, r, t, p, p.adj, ci_lower, ci_upper) and the merged data.
}
\description{
Perform correlation analysis of the mRNA/protein expression data in CPTAC database.
}
\examples{
\dontrun{
results <- cor_cancer_genelist(dataset = "GSE62113",
                               id1 = "STAT3",tumor_subtype = "LC",
                               id2 = c("TNS1", "TP53"),
                               sample_type = c("Tumor", "Normal"),
                               cor_method = "pearson",
                               adjust_method = "BH")
}
}
