% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gcas_drug.R
\name{cor_gcas_drug}
\alias{cor_gcas_drug}
\title{Correlation Analysis of Drug Sensitivity}
\usage{
cor_gcas_drug(df, cor_method = "pearson", Target.pathway = c("Cell cycle"))
}
\arguments{
\item{df}{The expression data of the target gene in multiple datasets, obtained by the get_expr_data() function.}

\item{cor_method}{Method for correlation analysis, default "pearson".}

\item{Target.pathway}{The signaling pathways of anti-tumor drug targets, default "Cell cycle". Use "drug_info" to get the detailed information of these drugs.}
}
\description{
Calculate the correlation between target gene expression and anti-tumor drug sensitivity in multiple datasets.
}
\examples{
\dontrun{
dataset <- c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210",
             "GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072",
             "GSE74706","GSE18842","GSE62113")
df <- get_expr_data(genes = "TNS1", datasets = dataset)
result <- cor_gcas_drug(df, Target.pathway = c("Cell cycle"))
}
}
