% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gcas_genelist.R
\name{cor_gcas_genelist}
\alias{cor_gcas_genelist}
\title{Perform correlation analysis}
\usage{
cor_gcas_genelist(
  df,
  geneset_data,
  tumor_subtype = NULL,
  sample_type = c("Tumor", "Normal"),
  cor_method = "pearson"
)
}
\arguments{
\item{df}{The expression data of the target gene in multiple datasets, obtained by the get_expr_data() function.}

\item{geneset_data}{The expression data of a genelist in multiple datasets, obtained by the get_expr_data() function.}

\item{tumor_subtype}{Tumor subtype used for correlation analysis, default is NULL.}

\item{sample_type}{Sample type used for correlation analysis, default all types: c("Tumor", "Normal").}

\item{cor_method}{Method for correlation analysis, default "pearson".}
}
\description{
Perform correlation analysis of the expression data in CPTAC database.
}
\examples{
\dontrun{
genelist <- c("SIRPA","CTLA4","TIGIT","LAG3","VSIR","LILRB2","SIGLEC7","HAVCR2","LILRB4","PDCD1","BTLA")
dataset <- c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210","GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072","GSE74706","GSE18842","GSE62113")
df <- get_expr_data(genes = "TNS1",datasets = dataset)
geneset_data <- get_expr_data(genes = genelist ,datasets = dataset)
result <- cor_gcas_genelist(df, geneset_data, sample_type = c("Tumor"))
}
}
