% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Summary Statistics for Gene Expression Data}
\usage{
data_summary(df, tumor_subtype = NULL, method = "t.test")
}
\arguments{
\item{df}{A dataframe containing gene expression data, with columns: 'dataset', 'subtype', and the gene expression values.}

\item{tumor_subtype}{A character string specifying the tumor subtype to be analyzed. Default is NULL, which means all tumor subtypes will be included.}

\item{method}{A character string specifying the method for hypothesis testing. Options are "t.test" for t-test and "wilcox" for Wilcoxon test. Default is "t.test".}
}
\value{
A dataframe with summary statistics and p-values for each dataset.
}
\description{
Compute summary statistics (mean, standard deviation, etc.) and perform hypothesis testing (t-test or Wilcoxon test) for gene expression data across different datasets.
}
\examples{
\dontrun{
df <- get_expr_data(datasets = c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210","GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072","GSE74706","GSE18842","GSE62113"), genes = "GAPDH")
results <- data_summary(df, tumor_subtype = "LUAD")
}
}
