% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_summary.R
\name{datasets_summary}
\alias{datasets_summary}
\title{Summarize Datasets by Tumor Subtype}
\usage{
datasets_summary(tumor_subtype = "NSCLC")
}
\arguments{
\item{tumor_subtype}{A character string specifying the tumor subtype to filter the datasets. Default is "NSCLC".}
}
\value{
A data frame summarizing the number of Normal, Tumor, and Premalignant samples, as well as the pairing status for each dataset.
}
\description{
This function summarizes the sample counts and pairing status for datasets based on a specified tumor subtype.
}
\examples{
\dontrun{
dt_sum <- datasets_summary("NSCLC")
}
}
