% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_analysis_utils.R
\name{.determine_sample_type}
\alias{.determine_sample_type}
\title{Internal utility function to determine sample type}
\usage{
.determine_sample_type(df, tumor_subtype = NULL, subtype_col = "subtype")
}
\arguments{
\item{df}{The dataframe containing sample data.}

\item{tumor_subtype}{The tumor subtypes to consider.}

\item{subtype_col}{The column name for subtype information.}
}
\value{
The dataframe with an added "type" column.
}
\description{
This is an internal helper function to determine sample types (Tumor/Normal).
}
\keyword{internal}
