% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_analysis_utils.R
\name{.extract_tumor_subtype}
\alias{.extract_tumor_subtype}
\title{Internal utility function to extract tumor subtypes}
\usage{
.extract_tumor_subtype(subtype_data, tumor_subtype)
}
\arguments{
\item{subtype_data}{The subtype mapping data.}

\item{tumor_subtype}{The tumor subtypes to extract.}
}
\value{
A vector of extracted subtypes.
}
\description{
This is an internal helper function to extract and validate tumor subtypes.
}
\keyword{internal}
