% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utils.R
\name{gcas_theme}
\alias{gcas_theme}
\title{GCAS ggplot2 theme}
\usage{
gcas_theme(base_size = 16, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
A consistent theme for all GCAS visualizations.
}
\examples{
\dontrun{
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  gcas_theme()
}
}
