% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get GEO expression data}
\usage{
get_data(table = "GSE74706", action = "expression", genes = c("GAPDH", "TNS1"))
}
\arguments{
\item{table}{For action = expression, use dataset$Abbre to get all tables; For action = clinic, remove _protein/_mRNA/_Phospho from dataset$Abbre.}

\item{action}{"expression", "sample_info" or "geo_data".}

\item{gene}{Gene symbols, you can input one or multiple symbols.}
}
\description{
Get the GEO data by using the api. All results saved in MySQL database.
}
\examples{
\dontrun{
results <- get_data(table = "GSE74706", action = "expression", genes = c("GAPDH","TNS1"))
}
}
