% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_data.R
\name{get_expr_data}
\alias{get_expr_data}
\title{Retrieve Gene Expression Data}
\usage{
get_expr_data(datasets, genes)
}
\arguments{
\item{datasets}{A character vector of dataset identifiers.}

\item{genes}{A character vector of gene identifiers.}
}
\value{
A dataframe containing expression data for the specified genes from the given datasets.
}
\description{
Retrieve expression data for specified genes from given datasets.
}
\examples{
\dontrun{
results <- get_expr_data(datasets = "GSE74706", genes = c("GAPDH","TNS1"))
results <- get_expr_data(datasets = c("GSE62113","GSE74706"), genes = "GAPDH")
results <- get_expr_data(datasets = c("GSE62113","GSE74706"), genes = c("SIRPA","CTLA4","TIGIT","LAG3","VSIR","LILRB2","SIGLEC7","HAVCR2","LILRB4","PDCD1","BTLA"))
}
}
