% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OSF_data.R
\name{get_OSF_data}
\alias{get_OSF_data}
\title{Get GEO expression dataset and sample information in OSF repository}
\usage{
get_OSF_data(table = "GSE19188", action = "geo_data", retries = 2)
}
\arguments{
\item{table}{A character string specifying the GEO dataset identifier (e.g., "GSE19188").}

\item{action}{A character string specifying the action, either "geo_data" to retrieve the expression data or "sample_info" to retrieve the sample information.}
}
\value{
A data frame containing the requested data.
}
\description{
Retrieve GEO expression datasets and sample information from the OSF repository.
}
\examples{
\dontrun{
df <- get_OSF_data(table = "GSE74706", action = "sample_info")
df2 <- get_OSF_data(table = "GSE74706", action = "geo_data")
}
}
