% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_analysis.R
\name{GSEA_analysis}
\alias{GSEA_analysis}
\title{Perform GSEA analysis for single gene}
\usage{
GSEA_analysis(data, gmt_file, pvalue_cutoff = 0.05, data_type = "correlation")
}
\arguments{
\item{data}{A data frame containing gene names and corresponding values. For correlation results, the columns should be named `gene` and `r`. For limma results, the columns should be named `gene` and `logFC`.}

\item{gmt_file}{Path to the GMT file containing gene sets, or directly pass GO/KEGG/Reactome datasets.}

\item{pvalue_cutoff}{Numeric, the p-value threshold for significance. Default is 0.05.}

\item{data_type}{Character, type of the input data. Either "correlation" for correlation analysis results or "limma" for limma differential analysis results.}
}
\value{
A GSEA analysis result object.
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) based on either correlation results or limma differential analysis results.
}
\examples{
\dontrun{
df <- get_OSF_data(table = "GSE74706", action = "geo_data")
results <- DEGs_analysis(df,tumor_subtype =c("NSCLC"))
gsea_result <- GSEA_analysis(results,  gmt_file = BP_GMT_7.5.1, data_type = "limma")

results <- coexpression_analysis(df,"RPN1")
gsea_result <- GSEA_analysis(results,  gmt_file = BP_GMT_7.5.1)
}
}
