% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{interpret_correlation}
\alias{interpret_correlation}
\title{Interpret Correlation Coefficient}
\usage{
interpret_correlation(r)
}
\arguments{
\item{r}{The correlation coefficient.}
}
\value{
A character string describing the correlation strength.
}
\description{
Provide interpretation of correlation coefficient magnitude.
}
