% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clinic_data.R
\name{merge_clinic_data}
\alias{merge_clinic_data}
\title{Merge sample_info data}
\usage{
merge_clinic_data(table = "GSE19188", data_input)
}
\arguments{
\item{table}{Character. The name of the dataset table to retrieve sample information from. Default is "GSE19188".}

\item{data_input}{Data frame. Expression data obtained from get_expr_data() function.}
}
\value{
Data frame. Merged data containing both expression data and sample information.
}
\description{
Get sample_info data and merge it with expression data.
}
\examples{
\dontrun{
data_input <- get_expr_data("GSE19188", "TP53")
results <- merge_clinic_data("GSE19188",data_input)
}
}
