% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{minimum_detectable_effect}
\alias{minimum_detectable_effect}
\title{Calculate Minimum Detectable Effect Size}
\usage{
minimum_detectable_effect(n1, n2 = NULL, power = 0.8, alpha = 0.05)
}
\arguments{
\item{n1}{Sample size in group 1.}

\item{n2}{Sample size in group 2. If NULL, assumed equal to n1.}

\item{power}{Desired statistical power. Default is 0.8.}

\item{alpha}{Significance level. Default is 0.05.}
}
\value{
The minimum detectable effect size (Cohen's d).
}
\description{
Calculate the minimum effect size detectable with given sample size and power.
}
\examples{
\dontrun{
min_effect <- minimum_detectable_effect(n1 = 50, n2 = 50)
}
}
