% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logFC_scatter.R
\name{plot_logFC_scatter}
\alias{plot_logFC_scatter}
\title{Plot logFC scatter}
\usage{
plot_logFC_scatter(
  results,
  p.cut = 0.05,
  logFC.cut = 1,
  colors = c("blue", "grey20", "red")
)
}
\arguments{
\item{results}{Data frame. The results data frame containing columns for dataset, gene, logFC, and p.value.}

\item{p.cut}{Numeric. The cutoff for adjusted p-value to determine significance. Default is 0.05.}

\item{logFC.cut}{Numeric. The cutoff for log fold change to determine significance. Default is 1.}

\item{colors}{A vector of color panel, default c("blue", "grey20", "red").}
}
\value{
A ggplot object representing the scatter.
}
\description{
This function generates a scatter of log fold changes (logFC) for a gene across different datasets.
It includes significance annotations based on p-values.
}
\examples{
\dontrun{
df <- get_expr_data(datasets = c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210","GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072","GSE74706","GSE18842","GSE62113"), genes = "GAPDH")
results <- data_summary(df, tumor_subtype = "LUAD")
scatter <- plot_logFC_scatter(results, logFC.cut = 0.5, colors = c("blue","grey20", "red"))
print(scatter)
}
}
