% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta_forest.R
\name{plot_meta_forest}
\alias{plot_meta_forest}
\title{Meta analysis in multiple datasets}
\usage{
plot_meta_forest(results, method = "wilcox", k.min = 10)
}
\arguments{
\item{results}{Data frame. The results data frame containing columns for dataset, n_Tumor, mean_Tumor, sd_Tumor, n_Normal, mean_Normal, and sd_Normal.}

\item{method}{Character. The statistical method to use (default is "wilcox").}

\item{k.min}{Integer. Minimum number of studies for bias test (default is 7).}

\item{cohort}{Data cohort, for example, "LUAD_APOLLO", "LUAD_CPTAC".}

\item{data_input}{Expression data obtained from get_expr_data() function.}
}
\value{
A forest plot object.
}
\description{
Plotting volcano plot for DEGs between tumor and normal samples in CPTAC datasets.

This function performs a meta-analysis on multiple datasets and generates a forest plot.
It also tests for publication bias.
}
\examples{
\dontrun{
df <- get_expr_data(datasets = c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210","GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072","GSE74706","GSE18842","GSE62113"), genes = "GAPDH")
results <- data_summary(df, tumor_subtype = "LUAD")
plot_meta_forest(results)
}
}
