% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plot Venn Diagram for Differentially Expressed Genes}
\usage{
plot_venn(results, fill_colors = NULL, palette = "Set1", lty = 2, ...)
}
\arguments{
\item{results}{List of character vectors. Each vector contains DEGs for a specific dataset.}

\item{fill_colors}{Character vector. Colors to fill the Venn diagram circles. Default is NULL, which uses a palette.}

\item{palette}{Character. Name of the RColorBrewer palette to use if fill_colors is not specified. Default is "Set1".}

\item{lty}{Numeric. Line type for the circles in the Venn diagram. Default is 2 (dashed line).}

\item{...}{Additional arguments passed to venn.diagram function.}
}
\value{
A list of intersected DEGs.
}
\description{
This function plots a Venn diagram for lists of differentially expressed genes (DEGs) across multiple datasets.
}
\examples{
\dontrun{
df1 <- get_OSF_data(table = "GSE31210", action = "geo_data")
results1 <- DEGs_analysis(df1)
df2 <- get_OSF_data(table = "GSE19188", action = "geo_data")
results2 <- DEGs_analysis(df2)
DEGs_lists <- list("GSE31210" = results1, "GSE19188" = results2)
results <- get_DEGs_list(DEGs_lists)
plot_venn(results$DEG_up, palette = "Set1")
plot_venn(results$DEG_up, fill_colors = c("red", "green", "blue"), alpha = 0.5, cex = 1.5)
}
}
