% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot for DEGs}
\usage{
plot_volcano(
  results,
  p.cut = 0.05,
  logFC.cut = 1,
  show.top = FALSE,
  show.labels = NULL,
  colors = c("blue", "grey20", "red")
)
}
\arguments{
\item{results}{DataFrame. The results from DEGs analysis containing columns 'adj.P.Val', 'P.Value', 'logFC', and 'gene'.}

\item{p.cut}{Numeric. The cutoff for adjusted p-value to determine significance. Default is 0.05.}

\item{logFC.cut}{Numeric. The cutoff for log fold change to determine significance. Default is 1.}

\item{show.top}{Logical. If TRUE, labels the top 5 up- and downregulated genes. Default is FALSE.}

\item{show.labels}{Character vector. Specific gene labels to show. Default is NULL.}

\item{colors}{A vector of color panel, default c("blue", "grey20", "red").}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
Plotting volcano plot for DEGs between tumor and normal samples in CPTAC datasets.
}
\examples{
\dontrun{
df <- get_OSF_data(table = "GSE74706", action = "geo_data")
results <- DEGs_analysis(df)
plot_volcano(results)
}
}
