% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRA_analysis.R
\name{RRA_analysis}
\alias{RRA_analysis}
\title{Perform Robust Rank Aggregation (RRA) analysis on DEGs lists}
\usage{
RRA_analysis(
  DEGs_lists,
  top.num = 0,
  rra.p = 0.05,
  logFC_cut = 1,
  p_cut = 0.05
)
}
\arguments{
\item{DEGs_lists}{A list of DEGs data frames. Each data frame should contain at least a 'gene' column and a 'logFC' column.}

\item{top.num}{Numeric, the number of top genes to select based on their ranks. Default is 0, which selects all genes passing the thresholds.}

\item{rra.p}{Numeric, the p-value threshold for RRA. Default is 0.05.}

\item{logFC_cut}{Numeric, the log fold change threshold for filtering genes. Default is 1.}

\item{p_cut}{Numeric, the p-value threshold for filtering genes. Default is 0.05.}
}
\value{
A list containing the number of up- and down-regulated genes and a data matrix of aggregated log fold changes.
}
\description{
This function performs RRA analysis on differentially expressed genes (DEGs) lists
obtained from various studies. It ranks genes based on their differential expression
and aggregates the ranks to identify consistently regulated genes across studies.
}
\examples{
\dontrun{
df1 <- get_OSF_data(table = "GSE31210", action = "geo_data")
results1 <- DEGs_analysis(df1)
df2 <- get_OSF_data(table = "GSE19188", action = "geo_data")
results2 <- DEGs_analysis(df2)
DEGs_lists <- list("GSE31210" = results1, "GSE19188" = results2)
RRA_results <- RRA_analysis(DEGs_lists)
ComplexHeatmap::pheatmap(RRA_results$RRA_results)
}
}
