% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{set_gcas_api_url}
\alias{set_gcas_api_url}
\title{Set GCAS API URL}
\usage{
set_gcas_api_url(url)
}
\arguments{
\item{url}{The new API URL.}
}
\description{
Set a custom GCAS API URL for the current session.
}
