% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_cor_heatmap.R
\name{viz_cor_heatmap}
\alias{viz_cor_heatmap}
\title{Visualization of correlation results}
\usage{
viz_cor_heatmap(r, p)
}
\arguments{
\item{r}{The correlation coefficient matrix r of the correlation analysis results obtained from the functions cor_pancancer_genelist(), cor_pancancer_TIL(), and cor_pancancer_drug().}

\item{p}{The P-value matrix p of the correlation analysis results obtained from the functions cor_pancancer_genelist(), cor_pancancer_TIL(), and cor_pancancer_drug().}
}
\description{
Presenting correlation analysis results using heat maps based on ggplot2.
}
\examples{
\dontrun{
genelist <- c("SIRPA","CTLA4","TIGIT","LAG3","VSIR","LILRB2","SIGLEC7","HAVCR2","LILRB4","PDCD1","BTLA")
dataset <-  c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210","GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072","GSE74706","GSE18842","GSE62113")
df <- get_expr_data(genes = "TNS1", datasets = dataset)
geneset_data <- get_expr_data(genes = genelist ,datasets = dataset)
result <- cor_gcas_genelist(df, geneset_data, sample_type = c("Tumor"))
viz_cor_heatmap(result$r, result$p)
}
}
