% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_cor_volcano.R
\name{viz_cor_volcano}
\alias{viz_cor_volcano}
\title{Volcano plot for the correlation analysis}
\usage{
viz_cor_volcano(
  cor_result,
  item,
  p.cut = 0.05,
  r.cut = 0.3,
  colors = c("blue", "grey20", "red")
)
}
\arguments{
\item{cor_result}{DataFrame. The results from cor_gcas_genelist analysis.}

\item{item}{Character. Gene symbol or drug name or immune cell.}

\item{p.cut}{Numeric. The cutoff for adjusted p-value to determine significance. Default is 0.05.}

\item{r.cut}{Numeric. The cutoff for log fold change to determine significance. Default is 0.3.}

\item{colors}{A vector of color panel, default c("blue", "grey20", "red").}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
Plotting volcano plot for the correlation analysis of a specific gene from the result  of cor_gcas_genelist().
}
\examples{
\dontrun{
genelist <- c("SIRPA","CTLA4","TIGIT","LAG3","VSIR","LILRB2","SIGLEC7","HAVCR2","LILRB4","PDCD1","BTLA")
dataset <- c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210","GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072","GSE74706","GSE18842","GSE62113")
df <- get_expr_data(genes = "TNS1",datasets = dataset)
geneset_data <- get_expr_data(genes = genelist ,datasets = dataset)
cor_result <- cor_gcas_genelist(df, geneset_data, sample_type = c("Tumor"))
viz_cor_volcano(cor_result, "LILRB4", p.cut = 0.5, r.cut = 0.1,colors = c("blue" ,"grey20", "red"))
}
}
