% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_TvsN.R
\name{viz_TvsN}
\alias{viz_TvsN}
\title{Visualizing genes expression}
\usage{
viz_TvsN(
  df,
  df_type = c("single", "multi_gene", "multi_set"),
  tumor_subtype = NULL,
  Show.P.value = TRUE,
  Show.P.label = TRUE,
  Method = "t.test",
  values = c("#00AFBB", "#FC4E07"),
  Show.n = TRUE,
  Show.n.location = "default"
)
}
\arguments{
\item{df}{Gene expression data obtained from get_expr_data().}

\item{df_type}{The type of gene expression data, one value of "single","multi_gene", and "multi_set".}

\item{Show.P.value}{Whether to display the results of differential analysis, default TRUE.}

\item{Show.P.label}{Whether to display significance markers for differential analysis, default TRUE.}

\item{Method}{Methods of differential analysis, "t.test" or "limma", default "t.test".}

\item{values}{Color palette for normal and tumor groups. Default c("#00AFBB", "#FC4E07").}

\item{Show.n}{Display sample size.}

\item{Show.n.location}{Y-axis position displayed for sample size.}
}
\description{
Visualizing the different expression of mRNA expression data between Tumor and Normal tissues in GEO database.
}
\examples{
\dontrun{
df_single <- get_expr_data(datasets = "GSE27262",genes = c("TP53"))
df_multi_gene <- get_expr_data(datasets = "GSE27262",genes = c("TP53","TNS1"))
df_multi_set <- get_expr_data(datasets = c("GSE27262","GSE7670","GSE19188","GSE19804","GSE30219","GSE31210","GSE32665","GSE32863","GSE43458","GSE46539","GSE75037","GSE10072","GSE74706","GSE18842","GSE62113"), genes = "GAPDH")
viz_TvsN(df_single,df_type = "single")
viz_TvsN(df_multi_gene,df_type = "multi_gene",tumor_subtype ="LC")
viz_TvsN(df_multi_set,df_type = "multi_set")
}
}
