\name{abbr_full}
\alias{abbr_full}
\docType{data}
\title{
The full name of tumor abbreviation
}
\description{
The full name of tumor abbreviation
}
\usage{data("abbr_full")}
\format{
  A data frame with 132 observations on the following 3 variables.
  \describe{
    \item{\code{Type}}{a character vector}
    \item{\code{Subtype}}{a character vector}
    \item{\code{Full_name}}{a character vector}
  }
}
\examples{
data(abbr_full)
str(abbr_full)
}
