% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gcas_TIL.R
\name{cor_gcas_TIL}
\alias{cor_gcas_TIL}
\title{Correlation analysis of immune cells infiltration}
\usage{
cor_gcas_TIL(df, cor_method = "spearman", TIL_type = "TIMER")
}
\arguments{
\item{df}{The expression data of the target gene in multiple datasets, obtained by the get_expr_data() function.}

\item{cor_method}{Method for correlation analysis, default "spearman".}

\item{TIL_type}{Algorithm for calculating immune cell infiltration, default "TIMER".}
}
\description{
Calculate the correlation between target gene expression and immune cells infiltration in multiple datasets.
}
\examples{
\dontrun{
dataset <- c("GSE27262", "GSE7670", "GSE19188", "GSE19804", "GSE30219",
             "GSE31210", "GSE32665", "GSE32863", "GSE43458", "GSE46539",
             "GSE75037", "GSE10072", "GSE74706", "GSE18842", "GSE62113")
df <- get_expr_data(genes = "TNS1", datasets = dataset)
result <- cor_gcas_TIL(df, cor_method = "spearman", TIL_type = "TIMER")
}
}
