<<<<<<< HEAD
\name{TIL_map}
\alias{TIL_map}
\docType{data}
\title{
Mapping of immune cell types and algorithms
}
\description{
Mapping of immune cell types and algorithms
}
\usage{data("TIL_map")}
\format{
  A data frame with 137 observations on the following 2 variables.
  \describe{
    \item{\code{cell_type}}{a character vector}
    \item{\code{algorithm}}{a character vector}
  }
}
\examples{
data(TIL_map)
str(TIL_map)
}
=======
\name{TIL_map}
\alias{TIL_map}
\docType{data}
\title{
Mapping of immune cell types and algorithms
}
\description{
Mapping of immune cell types and algorithms
}
\usage{data("TIL_map")}
\format{
  A data frame with 137 observations on the following 2 variables.
  \describe{
    \item{\code{cell_type}}{a character vector}
    \item{\code{algorithm}}{a character vector}
  }
}
\examples{
data(TIL_map)
str(TIL_map)
}
>>>>>>> 6f4eea48388fdad44a0cc783c245e2ab56dbb4df
